+++ lib/lightning_web/live/components/common.ex
@@ -6,6 +6,104 @@ defmodule LightningWeb.Components.Common do
 
   alias Phoenix.LiveView.JS
 
+  attr :id, :string, default: "alert"
+  attr :type, :string, required: true
+  attr :class, :string, default: ""
+  attr :header, :string, default: nil
+  slot :message
+  attr :link_right, :map, required: false, default: nil
+  attr :border, :boolean, default: false
+  attr :action_position, :atom, default: :right
+  attr :actions, :list, default: []
+
+  @doc """
+  Use an alert to convey critical information about the state of a specific
+  system or artifact (a credential, a job, a workflow, etc.) and provide a type
+  to indicate whether the alert is info, success, warning, or error.
+  """
+  def alert(assigns) do
+    color =
+      case assigns.type do
+        "success" -> "green"
+        "warning" -> "yellow"
+        "danger" -> "red"
+        _info -> "blue"
+      end
+
+    icon =
+      case assigns.type do
+        "success" -> "hero-check-circle-solid"
+        "warning" -> "hero-exclamation-triangle-solid"
+        "error" -> "hero-x-circle-solid"
+        _info -> "hero-information-circle-solid"
+      end
+
+    assigns =
+      assign(assigns,
+        color: color,
+        icon: icon,
+        class: assigns.class
+      )
+
+    ~H"""
+    <div id={@id} class={"rounded-md bg-#{@color}-50 p-4 #{@class}"}>
+      <div class="flex">
+        <div class="flex-shrink-0">
+          <.icon name={@icon} class={"mb-1 h-5 w-5 text-#{@color}-400"} />
+        </div>
+        <div class={[
+          "ml-3",
+          assigns[:link_right] && "flex-1 md:flex md:justify-between"
+        ]}>
+          <%= if @header do %>
+            <h3 class={"text-sm font-medium text-#{@color}-800"}><%= @header %></h3>
+            <div class={"mt-2 text-sm text-#{@color}-700"}>
+              <%= render_slot(@message) %>
+            </div>
+          <% else %>
+            <div class={"text-sm text-#{@color}-700"}>
+              <%= render_slot(@message) %>
+            </div>
+          <% end %>
+          <%= if assigns[:link_right] do %>
+            <p class="mt-3 text-sm md:ml-6 md:mt-0">
+              <a
+                href={@link_right.target}
+                class={"whitespace-nowrap font-medium text-#{@color}-700 hover:text-#{@color}-600"}
+              >
+                <%= @link_right.text %>
+                <span aria-hidden="true"> &rarr;</span>
+              </a>
+              <%= render_slot(@link_right) %>
+            </p>
+          <% else %>
+            <div :if={@actions} class={["mt-4"]}>
+              <div class="-mx-2 -my-1.5 flex">
+                <%= for item <- @actions do %>
+                  <%= case item do %>
+                    <% %{id: id, text: text, click: click, target: target} -> %>
+                      <button
+                        id={id}
+                        type="button"
+                        phx-click={click}
+                        phx-target={target}
+                        class={"rounded-md bg-#{@color}-50 px-2 py-1.5 text-sm font-medium text-#{@color}-800 hover:bg-#{@color}-100 focus:outline-none focus:ring-2 focus:ring-#{@color}-600 focus:ring-offset-2 focus:ring-offset-#{@color}-50"}
+                      >
+                        <%= text %>
+                      </button>
+                    <% element -> %>
+                      <%= element %>
+                  <% end %>
+                <% end %>
+              </div>
+            </div>
+          <% end %>
+        </div>
+      </div>
+    </div>
+    """
+  end
+
   attr :id, :string, required: true
   attr :version, :string, required: true
   attr :tooltip, :string, required: false