+++ test/lightning/credentials_test.exs
@@ -196,6 +196,35 @@ defmodule Lightning.CredentialsTest do
       )
     end
 
+    test "schedule_credential_deletion/1 revokes token for oauth credentials" do
+      oauth_credential =
+        insert(:credential,
+          name: "My Credential",
+          schema: "oauth",
+          body: %{
+            "access_token" => "super_secret_access_token_123",
+            "refresh_token" => "super_secret_refresh_token_123",
+            "expires_in" => 3000
+          },
+          user: build(:user),
+          oauth_client: build(:oauth_client)
+        )
+
+      refute oauth_credential.scheduled_deletion
+
+      expect(Lightning.AuthProviders.OauthHTTPClient.Mock, :call, fn
+        env, _opts
+        when env.method == :post and
+               env.url == "http://example.com/oauth2/revoke" ->
+          {:ok, %Tesla.Env{status: 200, body: Jason.encode!(%{})}}
+      end)
+
+      {:ok, oauth_credential} =
+        Credentials.schedule_credential_deletion(oauth_credential)
+
+      assert oauth_credential.scheduled_deletion
+    end
+
     test "cancel_scheduled_deletion/1 sets scheduled_deletion to nil for a given credential" do
       # Set up a credential with a scheduled_deletion date
       # 1 hour from now, truncated to seconds
@@ -371,6 +400,38 @@ defmodule Lightning.CredentialsTest do
   end
 
   describe "update_credential/2" do
+    test "updates an Oauth credential with new scopes" do
+      credential =
+        credential_fixture(
+          body: %{
+            "access_token" => "ya29.a0AWY7CknfkidjXaoDTuNi",
+            "expires_at" => 10_000,
+            "refresh_token" => "1//03dATMQTmE5NSCgYIARAAGAMSNwF",
+            "scope" => "email calendar chat"
+          },
+          schema: "oauth"
+        )
+
+      update_attrs = %{
+        body: %{
+          "access_token" => "ya29.a0AWY7CknfkidjXaoDTuNi",
+          "refresh_token" => "1//03dATMQTmE5NSCgYIARAAGAMSNwF",
+          "expires_at" => 10_000,
+          "scope" => "email calendar"
+        }
+      }
+
+      assert {:ok, %Credential{} = credential} =
+               Credentials.update_credential(credential, update_attrs)
+
+      assert credential.body == %{
+               "access_token" => "ya29.a0AWY7CknfkidjXaoDTuNi",
+               "expires_at" => 10000,
+               "refresh_token" => "1//03dATMQTmE5NSCgYIARAAGAMSNwF",
+               "scope" => "email calendar"
+             }
+    end
+
     test "succeeds with valid data and associating with new project" do
       user = insert(:user)